+++ CHANGES.rst
@@ -2,14 +2,42 @@ Changelog
 =========
 
 
-1.1.1 (unreleased)
+2.1.2 (unreleased)
 ------------------
 
 - [new] Content type 'Bando' is admitted by default in 'Documenti e dati' folder.
   [arsenico13]
+
+
+2.1.1 (2021-05-04)
+------------------
+
+- Fix upgrade-step.
+  [cekk]
+
+
+2.1.0 (2021-05-04)
+------------------
+
+- Handle multiple twitter accounts in twitter endpoint.
+  [cekk]
+- Change twitter token field in registry (now is managed with interface).
+  [cekk]
+
+
+2.0.0 (2021-04-30)
+------------------
+
+- Update with new settings values from design.plone.contenttypes (version 3.0.0).
+  [cekk]
 - [dev] Fix CI
   [arsenico13]
-
+- Add custom image scales
+  [nzambello]
+- **search-filters** endpoint now return also a list of searchable portal_types.
+  [cekk]
+- Disable some types from *types_not_searched*.
+  [cekk]
 
 1.1.0 (2021-03-24)
 ------------------